pyproject.toml
[project]
name = "knowledgeforge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate==1.12.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.12.1",
    "asttokens==3.0.1",
    "attrs==25.4.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "beautifulsoup4==4.14.3",
    "build==1.4.0",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "chromadb==1.4.1",
    "click==8.3.1",
    "coloredlogs==15.0.1",
    "colorlog==6.10.1",
    "comm==0.2.3",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.20",
    "decorator==5.2.1",
    "dill==0.4.1",
    "distro==1.9.0",
    "docling>=2.73.0",
    "docling-core>=2.64.0",
    "docling-ibm-models>=3.11.0",
    "docling-parse>=4.7.3",
    "durationpy==0.10",
    "et-xmlfile==2.0.0",
    "executing==2.2.1",
    "faker==40.4.0",
    "fastapi==0.128.3",
    "filelock==3.20.3",
    "filetype==1.2.0",
    "flatbuffers==25.12.19",
    "frozenlist==1.8.0",
    "fsspec==2026.2.0",
    "googleapis-common-protos==1.72.0",
    "greenlet==3.3.1",
    "grpcio==1.78.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.36.2",
    "humanfriendly==10.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "importlib-resources==6.5.2",
    "iniconfig==2.3.0",
    "ipykernel==7.2.0",
    "ipython==9.10.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "jsonlines==4.0.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonref==1.1.0",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.8.0",
    "jupyter-core==5.9.1",
    "kubernetes==35.0.0",
    "langchain==1.2.9",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.9",
    "langchain-huggingface==1.2.0",
    "langchain-text-splitters==1.1.0",
    "langgraph==1.0.8",
    "langgraph-checkpoint==4.0.0",
    "langgraph-prebuilt==1.0.7",
    "langgraph-sdk==0.3.4",
    "langsmith==0.6.9",
    "latex2mathml==3.78.1",
    "lxml==6.0.2",
    "markdown-it-py==4.0.0",
    "marko==2.2.2",
    "markupsafe==3.0.3",
    "marshmallow==3.26.2",
    "matplotlib-inline==0.2.1",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "mpire==2.10.2",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "multiprocess==0.70.19",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.6.1",
    "numpy==2.4.2",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-cusparselt-cu12==0.6.2",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "oauthlib==3.3.1",
    "omegaconf==2.3.0",
    "onnxruntime==1.23.2",
    "opencv-python==4.13.0.92",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.39.1",
    "opentelemetry-exporter-otlp-proto-common==1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.39.1",
    "opentelemetry-proto==1.39.1",
    "opentelemetry-sdk==1.39.1",
    "opentelemetry-semantic-conventions==0.60b1",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "overrides==7.7.0",
    "packaging==26.0",
    "pandas==2.3.3",
    "parso==0.8.5",
    "pexpect==4.9.0",
    "pillow>=11.3.0",
    "pip==25.3",
    "platformdirs==4.5.1",
    "pluggy==1.6.0",
    "polyfactory==3.2.0",
    "posthog==5.4.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "protobuf==6.33.5",
    "psutil==7.2.2",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pybase64==1.4.3",
    "pyclipper==1.4.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pylatexenc==2.10",
    "pypdfium2==5.3.0",
    "pypika==0.51.1",
    "pyproject-hooks==1.2.0",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.2.0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.22",
    "python-pptx==1.0.2",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "rapidocr==3.6.0",
    "referencing==0.37.0",
    "regex==2026.1.15",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.3.2",
    "rpds-py==0.30.0",
    "rtree==1.4.1",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.0",
    "semchunk==2.2.2",
    "sentence-transformers==5.2.2",
    "setuptools==78.1.1",
    "shapely==2.1.2",
    "shellingham==1.5.4",
    "six==1.17.0",
    "soupsieve==2.8.3",
    "sqlalchemy==2.0.46",
    "stack-data==0.6.3",
    "starlette==0.52.1",
    "sympy==1.13.1",
    "tabulate==0.9.0",
    "tenacity==9.1.3",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "tokenizers==0.22.2",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "tornado==6.5.1",
    "tqdm==4.67.3",
    "traitlets==5.14.3",
    "transformers==4.57.6",
    "tree-sitter==0.25.2",
    "tree-sitter-c==0.24.1",
    "tree-sitter-javascript==0.25.0",
    "tree-sitter-python==0.25.0",
    "tree-sitter-typescript==0.23.2",
    "triton==3.2.0",
    "typer==0.21.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "uuid-utils==0.14.0",
    "uvicorn==0.40.0",
    "uvloop==0.22.1",
    "watchdog==6.0.0",
    "watchfiles==1.1.1",
    "wcwidth==0.5.3",
    "websocket-client==1.9.0",
    "websockets==16.0",
    "wheel==0.45.1",
    "xlsxwriter==3.2.9",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
markers = [
    "e2e: end-to-end integration tests (slow, loads ML models)",
]
asyncio_mode = "strict"
